pyproject.toml

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "odgs"
version = "3.3.0"
description = "Open Data Governance Standard — Sovereign Reference Implementation"
readme = "README.md"
requires-python = ">=3.9"
license = {text = "Apache-2.0"}
authors = [{name = "Metric Provenance", email = "partner@metricprovenance.com"}]
keywords = ["data-governance", "eu-ai-act", "sovereign", "compliance", "ai-safety"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: Apache Software License",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Software Development :: Libraries :: Application Frameworks",
]

dependencies = [
    "simpleeval==0.9.13",
    "typer>=0.9.0",
    "rich>=13.0.0",
    "uvicorn>=0.20.0",
    "fastapi>=0.95.0",
    "pydantic>=2.0.0",
    "python-dotenv>=1.0.0",
    "GitPython>=3.1.0",
    "certifi>=2024.2.2",
]

[project.optional-dependencies]
demo = ["streamlit>=1.30.0", "streamlit-agraph>=0.0.45", "pandas>=2.0.0"]
ai = ["google-genai>=1.0.0", "sse-starlette>=1.0.0"]
all = ["odgs[demo,ai]"]

[project.urls]
Homepage = "https://metricprovenance.com"
Repository = "https://github.com/MetricProvenance/odgs-protocol"
Documentation = "https://github.com/MetricProvenance/odgs-protocol/tree/main/docs"

[project.scripts]
odgs = "odgs.cli:app"

[tool.hatch.build.targets.wheel]
packages = ["2_INFORMATIVE_REFERENCE/src/odgs"]
exclude = [
    "2_INFORMATIVE_REFERENCE/src/odgs/ui/",
    "2_INFORMATIVE_REFERENCE/src/odgs/factory/",
    "2_INFORMATIVE_REFERENCE/src/odgs/audit_logs/",
    "2_INFORMATIVE_REFERENCE/src/odgs/sovereign_audit.log",
]

[tool.hatch.build.targets.wheel.sources]
"2_INFORMATIVE_REFERENCE/src" = ""

[tool.hatch.build.targets.sdist]
exclude = [
    "2_INFORMATIVE_REFERENCE/src/odgs/ui/",
    "2_INFORMATIVE_REFERENCE/src/odgs/factory/",
    "2_INFORMATIVE_REFERENCE/src/odgs/audit_logs/",
    ".env",
    "deploy/",
    "docker-compose.yml",
    "Dockerfile",
    ".streamlit/",
    "docs/internal/",
    "docs/images/",
    "COMMERCIALIZATION_PLAN.md",
    "PACKAGE_RELEASE_AUDIT.md",
    "OPENSOURCE_RELEASE_CLEANUP.md",
]
